+++ R/model.R
@@ -570,7 +570,7 @@ naomi_model_frame <- function(area_merged,
                       spec_unaware_untreated_prop_t4 = unaware_untreated_prop,
                       asfr_t4 = asfr,
                       frr_plhiv_t4 = frr_plhiv,
-                      frr_already_art_t4 = frr_already_art                      
+                      frr_already_art_t4 = frr_already_art
                     ),
              by = c("spectrum_region_code", "sex", "age_group")
            ) %>%
@@ -881,65 +881,6 @@ select_naomi_data <- function(
 
   stopifnot(methods::is(naomi_mf, "naomi_mf"))
 
-  ## Check anc_testing and art_number against Spectrum inputs.
-  ## Return NA if spec_program_data not provided
-  anc_testing_spectrum_aligned <- NA
-  art_number_spectrum_aligned <- NA
-
-  if (!is.null(spec_program_data)) {
-    stopifnot(methods::is(spec_program_data, "spec_program_data"))
-
-    if (!is.null(anc_testing)) {
-
-      anc_merged <- anc_testing %>%
-        dplyr::left_join(
-          dplyr::select(naomi_mf$mf_areas, area_id, spectrum_region_code),
-          by = "area_id"
-        ) %>%
-        tidyr::pivot_longer(dplyr::starts_with("anc"),
-                            names_to = "indicator",
-                            values_to = "value_naomi") %>%
-        dplyr::count(spectrum_region_code, year, indicator,
-                     wt = value_naomi, name = "value_naomi") %>%
-        dplyr::inner_join(
-          spec_program_data$anc_testing %>%
-          dplyr::rename("value_spectrum" = "value"),
-          by = c("spectrum_region_code", "indicator", "year")
-        )
-
-      anc_testing_spectrum_aligned <- all(anc_merged$value_naomi == anc_merged$value_spectrum)
-
-    } else {
-      ## If no ANC testing data, return TRUE
-      anc_testing_spectrum_aligned <- TRUE
-    }
-
-    if (!is.null(art_number)) {
-
-      art_merged <- art_number %>%
-        dplyr::left_join(
-          dplyr::select(naomi_mf$mf_areas, area_id, spectrum_region_code),
-          by = "area_id"
-        ) %>%
-        dplyr::count(spectrum_region_code, sex, age_group, calendar_quarter,
-                     wt = art_current, name = "art_current_naomi") %>%
-        dplyr::inner_join(
-          spec_program_data$art_dec31 %>%
-          dplyr::mutate(
-            calendar_quarter = paste0("CY", year, "Q4"),
-            year = NULL
-          ),
-          by = c("spectrum_region_code", "sex", "age_group", "calendar_quarter")
-        )
-
-      art_number_spectrum_aligned <- all(art_merged$art_current_naomi == art_merged$art_dec31)
-
-    } else {
-      ## If no ANC testing data, return TRUE
-      art_number_spectrum_aligned <- TRUE
-    }
-  }
-
   common_surveys <- intersect(artcov_survey_ids, vls_survey_ids)
   if (length(common_surveys)) {
     stop(t_("ART_COV_AND_VLS_SAME_SURVEY",
@@ -1101,9 +1042,7 @@ select_naomi_data <- function(
                        artnum_calendar_quarter_t1 = artnum_calendar_quarter_t1,
                        artnum_calendar_quarter_t2 = artnum_calendar_quarter_t2,
                        anc_prev_year_t1 = anc_artcov_year_t1,
-                       anc_prev_year_t2 = anc_artcov_year_t2,
-                       art_number_spectrum_aligned = art_number_spectrum_aligned,
-                       anc_testing_spectrum_aligned = anc_testing_spectrum_aligned)
+                       anc_prev_year_t2 = anc_artcov_year_t2)
 
   naomi_mf$data_options <- data_options