+++ R/pepfar-datapack.R
@@ -31,11 +31,18 @@ write_datapack_csv <- function(naomi_output,
                                psnu_level = NULL,
                                dmppt2_output = NULL) {
 
-  stopifnot(inherits(naomi_output, "naomi_output"))
-
   if (!grepl("\\.csv$", path, ignore.case = TRUE)) {
     path <- paste0(path, ".csv")
   }
+
+  datapack <- build_datapack_output(naomi_output, psnu_level, dmppt2_output)
+  naomi_write_csv(datapack, path)
+
+  path
+}
+
+build_datapack_output <- function(naomi_output, psnu_level, dmppt2_output) {
+  stopifnot(inherits(naomi_output, "naomi_output"))
 
   datapack_indicator_map <- naomi_read_csv(system_file("datapack", "datapack_indicator_mapping.csv"))
   datapack_age_group_map <- naomi_read_csv(system_file("datapack", "datapack_age_group_mapping.csv"))
@@ -77,7 +84,7 @@ write_datapack_csv <- function(naomi_output,
     dplyr::rename(
       indicator_code = datapack_indicator_code,
       dataelement_uid = datapack_indicator_id,
-      ) %>%
+    ) %>%
     dplyr::select(indicator, indicator_code, dataelement_uid, is_integer, calendar_quarter)
 
 
@@ -133,10 +140,10 @@ write_datapack_csv <- function(naomi_output,
       by = c("indicator", "calendar_quarter")
     ) %>%
     dplyr::filter(
-    (sex_naomi %in% datapack_sex_map$sex_naomi &
-       age_group %in% datapack_age_group_map$age_group |
-       sex_naomi == "both" & age_group == "Y000_999" & !anc_indicator |
-       sex_naomi == "female" & age_group == "Y015_049" & anc_indicator )
+      (sex_naomi %in% datapack_sex_map$sex_naomi &
+         age_group %in% datapack_age_group_map$age_group |
+         sex_naomi == "both" & age_group == "Y000_999" & !anc_indicator |
+         sex_naomi == "female" & age_group == "Y015_049" & anc_indicator )
     ) %>%
     dplyr::transmute(
       area_id,
@@ -185,7 +192,7 @@ write_datapack_csv <- function(naomi_output,
   ## Recode area_level as "Country" or "PSNU"
   dat$country_or_psnu <- ifelse(dat$area_level == 0, "Country", "PSNU")
 
-  datapack <- dat %>%
+  dat %>%
     dplyr::select(
       psnu,
       psnu_uid,
@@ -203,10 +210,43 @@ write_datapack_csv <- function(naomi_output,
       country_or_psnu
      
     )
+}
 
-  naomi_write_csv(datapack, path)
+build_datapack_metadata <- function(naomi_output, ids) {
+  cqs <- c(naomi_output$fit$model_options$calendar_quarter_t1,
+           naomi_output$fit$model_options$calendar_quarter_t2,
+           naomi_output$fit$model_options$calendar_quarter_t3,
+           naomi_output$fit$model_options$calendar_quarter_t4,
+           naomi_output$fit$model_options$calendar_quarter_t5)
+  meta_period <- data.frame(
+    c("Time point", "t1", "t2", "t3", "t4", "t5"), c("Quarter", cqs)
+  )
 
-  path
+  info <- attr(naomi_output, "info")
+  inputs <- read.csv(text = info$inputs.csv, header = FALSE)
+
+  version <- data.frame("Naomi Version", utils::packageVersion("naomi"))
+
+  if (!is.null(ids)) {
+    all_data <- list(version, ids, inputs, meta_period)
+  } else {
+    all_data <- list(version, inputs, meta_period)
+  }
+
+  max_cols <- max(vapply(all_data, ncol, numeric(1)))
+  col_names <- vapply(seq_len(max_cols), function(i) paste0("V", i), character(1))
+  empty_row <- data.frame(matrix("", ncol = max_cols, nrow = 1))
+  colnames(empty_row) <- col_names
+  all_data <- lapply(all_data, function(df) {
+    colnames(df) <- col_names[seq(1, ncol(df))]
+    if (ncol(df) < max_cols) {
+      df[, col_names[seq(ncol(df) + 1, max_cols)]] <- ""
+    }
+    df[] <- lapply(df, as.character)
+    rbind.data.frame(df, empty_row)
+  })
+
+  do.call(rbind.data.frame, all_data)
 }